refactor: simplify app names
The "learning_" prefix isn't really necessary for namespace collision,
since it's possible to use AppConfigs to change the package label in
edx-platform later if necessary.

Also changed the app names to their simple verb forms. For instance,
"learning_publishing" became publish.
