+++ openedx_learning/apps/core/publish/apps.py
@@ -5,10 +5,10 @@ publishing Django application initialization.
 from django.apps import AppConfig
 
 
-class PublishingConfig(AppConfig):
+class PublishConfig(AppConfig):
     """
     Configuration for the publishing Django application.
     """
-    name = "openedx_learning.apps.core.learning_publishing"
-    verbose_name = "Learning Core: Publishing"
+    name = "openedx_learning.apps.core.publish"
+    verbose_name = "Learning Core: Publish"
     default_auto_field = 'django.db.models.BigAutoField'
